Add utility to fill shapes with random values
